requirements/base.txt
# Flask
Flask==2.3

# database
SQLAlchemy==2.0
psycopg2-binary==2.9.7
alembic==1.12.0

# environments
#environs==9.5.0
python-dotenv==1.0.0
